# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "School Management Project",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/students/": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "Read Students",
        "description": "Retrieve students.",
        "operationId": "read_students",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentsPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "students"
        ],
        "summary": "Create Student",
        "description": "Create new student.",
        "operationId": "create_student",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/students/{id}": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "Read Student",
        "description": "Get student by ID.",
        "operationId": "read_student",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "students"
        ],
        "summary": "Update Student",
        "description": "Update a student.",
        "operationId": "update_student",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "students"
        ],
        "summary": "Delete Student",
        "description": "Delete a student.",
        "operationId": "delete_student",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Message": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "Message"
      },
...
      "StudentCreate": {
        "properties": {
          "first_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "First Name"
          },
          "middle_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Middle Name"
          },
          "last_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Last Name"
          },
          "contact": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact"
          },
          "form_id": {
            "type": "string",
            "format": "uuid",
            "title": "Form Id"
          },
          "fees": {
            "type": "number",
            "title": "Fees",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "first_name",
          "middle_name",
          "last_name",
          "contact",
          "form_id"
        ],
        "title": "StudentCreate"
      },
      "StudentPublic": {
        "properties": {
          "first_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "First Name"
          },
          "middle_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Middle Name"
          },
          "last_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Last Name"
          },
          "contact": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact"
          },
          "form_id": {
            "type": "string",
            "format": "uuid",
            "title": "Form Id"
          },
          "fees": {
            "type": "number",
            "title": "Fees",
            "default": 0
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          }
        },
        "type": "object",
        "required": [
          "first_name",
          "middle_name",
          "last_name",
          "contact",
          "form_id",
          "id",
          "owner_id"
        ],
        "title": "StudentPublic"
      },
      "StudentUpdate": {
        "properties": {
          "first_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "First Name"
          },
          "middle_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Middle Name"
          },
          "last_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Last Name"
          },
          "contact": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact"
          },
          "form_id": {
            "type": "string",
            "format": "uuid",
            "title": "Form Id"
          },
          "fees": {
            "type": "number",
            "title": "Fees",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "middle_name",
          "contact"
        ],
        "title": "StudentUpdate"
      },
      "StudentsPublic": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/StudentPublic"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "StudentsPublic"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/login/access-token"
          }
        }
      }
    }
  }
}